Inline AbstractDirectoryWalker
